+++ src/ui/components/noUsers/NoUsers.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { getImageUrl } from "../../../../utils";
+import { getImageUrl } from "../../../utils";
 
 import "./NoUsers.css";
